claude-4-deep-research/pyproject.toml
[project]
name = "claude-4-deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.0",
    "firecrawl-py>=2.7.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]

[dependency-groups]
notebook = [
    "ipykernel>=6.29.5",
]
